requirements.txt
aiofiles
aiohttp
docarray
docx2txt
fastapi
httpx
kubernetes
langchain
langchain-community
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-sdk
Pillow
prometheus-fastapi-instrumentator
pyjwt
pypdf
python-multipart
pyyaml
requests
shortuuid
uvicorn
